requirements.txt
langchain==0.3.25
openai==1.68.2
faiss-cpu==1.11.0
tiktoken==0.9.0
python-dotenv==1.1.0
chromadb==1.0.12
tavily-python==0.7.5
streamlit==1.45.1
